add -p to readme and fix install script

+++ install.sh
@@ -6,4 +6,8 @@ NAME="pamde"
 INSTALL_LOC="/usr/local/bin"
 
 sudo cp "./$NAME.sh" "$INSTALL_LOC/$NAME.sh"
-sudo ln -s "$INSTALL_LOC/$NAME.sh" "$INSTALL_LOC/$NAME"
+
+if test ! -e "$INSTALL_LOC/$NAME";
+then
+	sudo ln -s "$INSTALL_LOC/$NAME.sh" "$INSTALL_LOC/$NAME"
+fi
